requirements.txt
typer
# Include all your requirements here
typer==0.9.0
numpy==1.24.3
PyYAML==6.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch>1.12
acvl-utils>=0.2
dynamic-network-architectures>=0.2
tqdm    
dicom2nifti      
medpy    
scipy    
batchgenerators>=0.25    
numpy    
scikit-learn    
scikit-image  
SimpleITK>=2.2.1    
pandas    
graphviz    
tifffile
requests
nibabel    
matplotlib>=3.7.2 
seaborn    
imagecodecs    
yacs
kiwisolver
Pillow
connected-components-3d>=3.12.2
shapely
opencv-python